[Chore] #147- addPadding extension 위치 변경
