+++ Peekabook/Peekabook/Global/Extension/UITextField+.swift
@@ -14,5 +14,17 @@ extension UITextField {
         let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: -14, height: -10))
         self.leftView = paddingView
         self.leftViewMode = ViewMode.always
-  }
+    }
+    
+    func addLeftPadding(width: CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+    
+    func addRightPadding(width: CGFloat) {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
+        self.rightView = paddingView
+        self.rightViewMode = .always
+    }
 }
