--- Peekabook/Peekabook/Global/Extension/addPadding+UITextField.swift
@@ -1,22 +1,0 @@
-//
-//  addPadding+UITextField.swift
-//  Peekabook
-//
-//  Created by devxsby on 2022/12/31.
-//
-
-import UIKit
-
-extension UITextField {
-    func addLeftPadding(width: CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
-        self.leftView = paddingView
-        self.leftViewMode = .always
-    }
-    
-    func addRightPadding(width: CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
-        self.rightView = paddingView
-        self.rightViewMode = .always
-    }
-}
